+++ docs/source/CHANGELOG.md
@@ -19,3 +19,4 @@
 - Refactor views to enable compatibility with SQLAlchemy 2.0.39
 - Allow the `Factory` to accept database interface parameters
 - Use type annotations for declarative mappings
+- Accommodate systems with pre-defined configurations when testing